+++ llmexer/base/search.py
@@ -1,0 +1,1 @@
+"""Base methods and feature to be used in search CLI command."""